moved setupext into accelerate folder
